FEAT: Implement newsletter open tracking
This commit introduces the ability to track newsletter opens.

It adds an `opened` boolean and `opened_at` timestamp to the `campaign_subscriber` pivot table to record when a subscriber opens a newsletter.  A `NewsletterMail` mailable is also created.  These changes allow for tracking newsletter engagement and provide data for future analysis.